% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthcheck.R, R/llm_client.R
\name{llm_healthcheck}
\alias{llm_healthcheck}
\title{llm_healthcheck}
\usage{
llm_healthcheck()

llm_healthcheck()
}
\value{
list(ok=TRUE/FALSE, note=...)
}
\description{
Fonction pour vérifier si bien connecté au LLM
}
