% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mml_provider.R
\name{mml_get_items}
\alias{mml_get_items}
\title{Récupérer plusieurs items par handle/URL}
\usage{
mml_get_items(ids)
}
\arguments{
\item{ids}{vecteur de handles ou d'URLs}
}
\value{
list(items = list(...), fetched_at = Sys.time())
}
\description{
Récupérer plusieurs items par handle/URL
}
