% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mml_provider.R
\name{mml_price_by_url}
\alias{mml_price_by_url}
\title{Prix & infos d'un produit à partir d'une URL (ou handle)}
\usage{
mml_price_by_url(product_url)
}
\arguments{
\item{product_url}{URL produit complète (ou handle)}
}
\value{
list(title, price=list(amount,currency), url, image, source)
}
\description{
Prix & infos d'un produit à partir d'une URL (ou handle)
}
