% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mml_provider.R
\name{mml_product_js}
\alias{mml_product_js}
\title{Récupérer le JSON Shopify d'un produit par handle (product.js)}
\usage{
mml_product_js(handle)
}
\arguments{
\item{handle}{Handle du produit (ou URL produit)}
}
\value{
liste JSON (ou NULL si échec)
}
\description{
Récupérer le JSON Shopify d'un produit par handle (product.js)
}
