% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider.R
\name{provider_get_items}
\alias{provider_get_items}
\title{provider_get_items}
\usage{
provider_get_items(ids)
}
\arguments{
\item{ids}{}
}
\description{
Route une récupération par identifiants (ASIN / SKU / etc.)
}
