% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider.R
\name{provider_search}
\alias{provider_search}
\title{Recherche produits selon provider}
\usage{
provider_search(q, page = 1)
}
\arguments{
\item{q}{requête}

\item{page}{page (amazon)}
}
\description{
Recherche produits selon provider
}
