% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderers.R
\name{render_search}
\alias{render_search}
\title{Rendu HTML d'une recherche (liste de résultats)}
\usage{
render_search(res)
}
\description{
Rendu HTML d'une recherche (liste de résultats)
}
\keyword{internal}
