% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shopify-provider.R
\name{shopify_get_items}
\alias{shopify_get_items}
\title{shopify_get_items}
\usage{
shopify_get_items(ids)
}
\arguments{
\item{ids}{}
}
\description{
Récupération par IDs : pour ce MVP, on re-cherche par ID/texte
}
