% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shopify-provider.R
\name{shopify_search}
\alias{shopify_search}
\title{shopify_search}
\usage{
shopify_search(
  q,
  page = 1,
  after = NULL,
  country = "FR",
  language = "FR",
  limit = 10
)
}
\arguments{
\item{limit}{}
}
\description{
Recherche produits (query + context obligatoires ; country/lang/limit optionnels)
}
