% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider.R
\name{provider_get_items}
\alias{provider_get_items}
\title{Items par identifiants/URLs selon provider}
\usage{
provider_get_items(ids)
}
\arguments{
\item{ids}{vecteur}
}
\description{
Items par identifiants/URLs selon provider
}
